Split SingleRuleResult into separate file
